100 HTML files (batch 7)
The 7th batch of HTML files of separate developer log entries. I accidentally uploaded all 100 files in Batch 6 to the wrong directory, the 2nd time I have done this in the past 16 hours. I am not as mad, as I know it is fixable, and it also gives me more of an opportunity to use the method of mass-moving files once I learn how to do that in GitHub. There is still over 400 files left after this, but I am making major progress. Still no weekly entries have been uploaded yet.